requirements.txt
matplotlib==3.10.3
numpy==2.2.6
opencv-python==4.12.0.88
pandas==2.3.1
pillow==11.3.0
pycocotools==2.0.10
PyYAML==6.0.2
scipy==1.15.3
tensorboard==2.20.0
torch==2.7.1
torchmetrics==1.8.0
torchvision==0.22.1
torchviz==0.0.3
tqdm==4.67.1
typing_extensions==4.14.1